build scripts (#61)
add scripts for build

BuildTools is applied for our build. The same as SignalR used for build.
Give the developers consistent building experience.

All scripts come from:
https://github.com/aspnet/BuildTools/tree/dev/scripts/bootstrapper
It is very easy to build on both Windows and Linux:

build.cmd or build.sh

They will install the required dotnet in your system. It will try to
find the *.sln to build and run test defaultly.