Don't reuse `Utf8JsonWriter` with `Stream` output (#1504)
The `Utf8JsonWriter.Reset(Stream)` and  `Utf8JsonWriter.Reset()` don't release the underlying array buffer, which is a memory leak. The problem gets worse with the using of thread-pool. If a `Utf8JsonWriter` is initialized or reset with an `IBufferWriter<byte>`, that won't cause memory leak as the underlying array buffer is null in that case.

* Preserve `ReusableUtf8JsonWriter` for persistent mode.