Remove `InternalsVisibleTo` Functions extensions (#1637)
InternalsVisibleTo Functions extensions turns out to be problematic because it is easy to remove an internal interface depended by functions extensions and causes breaking changes. As the `WithEndpoints` interface is stable enough and we can make it public.

This change itself is breaking, but I think it is worth breaking once and then we can avoid the problem forever.