Refactor: use `HttpMessageHandler` to add access token
Motivation: In `ListConnectionsInGroup` API, the next page is referenced by an URI, and we need to generate access token for that URI. However, currently the access token generation is tightly coupled with `RestApiProvider`, which generates access token for fixed endpoint.

This PR moves the logic of handling access token out of `RestApiProvider` and `RestClient`. Instead, add a `HttpMessageHandler` to set `Authorization` header.